niigata: fix sized queue calculations
In the following situation:
- a pallet is ready to be unloaded
- the queue is currently full
- we just unheld a pallet to pick up and load from the queue
- the recently unheld pallet is still in the buffer and has not yet moved to the load station

We would also unhold the pallet ready to be unloaded.  Note the situation is two pallets in the
stocker are unheld (one empty and going to load and a second full going to be unloaded).
Normally, typically this isn't a problem because the priorities would cause the empty pallet
to move first.  But just in case, we now wait to unhold the full pallet to be unloaded
until the material is fully loaded.  (This also helps in case the operator cancels the
load for some reason.)