server: restrict load-unload events
The load-unload events have been restricted to a specific format
informally for a while, but this commit reflects it in the API.
Each face can have only one load and all material on a face must
be loaded together.  In particular, all material loaded on a face
must share the same elapsed time.