server: store workorders with potentially separate scheduleIds
In the old design, when workorders were updated they were replaced by
looking up the most recent scheduleid and replacing them.
Now, new jobs can be added when the list of jobs is empty and instead just workorders
are included.  The loading of workorders is then updated to always load the most
recent workorder snapshot.