mazak: switch proxy to symbolic links
This is so all the code is included in the main mazak nuget and the resulting folder will build.  I thought about moving the DB code into its own project instead of symbolic links, but then we need a new nuget to upload to nuget.org, a new reference to manage.