niigata: first attempt at test for bug
The bug is that archived jobs don't look at queued material to decide to prevent archiving, so a job is unarchived only to be immedietly archived.  The fix is a single change in BuildCellStatus, but I don't want to make it until there is a failing test for it.

This is the first attempt at such a failing test, but it causes others to fail as well so the test suite needs more updates.