mazak: update logging to support assigning castings at the end of load
Also, factor out the decision on syncing queues to a single function.  Right now, the syncing always occurs if there is an input queue, so the logging changes should never be hit (since material is already assigned by the time the load completes).  But, the logging changes are there to support in the future a more complicated logic in if a schedule and a queue should be synced.